requirements.txt
selenium
argparse
PyCryptodomex
tabulate
colorama
termcolor